% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{update_isorunCSIA}
\alias{update_isorunCSIA}
\title{install the version of isorun CSIA from GitHub}
\usage{
update_isorunCSIA(ref = "master")
}
\arguments{
\item{ref}{which version to install, master (=newest) is the default}
}
\description{
install the version of isorun CSIA from GitHub
}

